Move concept.h into core
