+++ include/ozo/core/nullable.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ozo/concept.h>
+#include <ozo/core/concept.h>
 #include <type_traits>
 #include <memory>
