+++ include/ozo/detail/typed_buffer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ozo/concept.h>
+#include <ozo/core/concept.h>
 
 namespace ozo {
 namespace detail {
