+++ include/ozo/send.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ozo/concept.h>
+#include <ozo/core/concept.h>
 #include <ozo/type_traits.h>
 #include <ozo/size_of.h>
 #include <ozo/ostream.h>
